Loading janky animations

+++ projects/atLib/source/IO/atScan.cpp
@@ -138,7 +138,7 @@ static int64_t _ScanInteger(const char *str, int64_t *pLen, const char *candidat
   const int64_t intlen = lastNum - nextChar;
   const int64_t baseChar = candidates[0];
 
-  int64_t mag = (int64_t)pow(nCandidates, intlen - 1);
+  int64_t mag = (int64_t)pow((double)nCandidates, double(intlen - 1));
   int64_t res = 0;
   for (int64_t c = nextChar; c < lastNum; ++c, mag /= nCandidates)
     res += (str[c] - baseChar) * mag;
